processor/illegal-parking/requirements.txt
pulsar-client==2.4.1
opencv-python
argparse
numpy
mxnet
influxdb